dev: rearrange code
